# paste-it-cleaned-app/swagger/swagger.json
{
    "swagger": "2.0",
    "info": { "version": "v1", "title": "PasteItCleaned Backend" },
    "paths": {
...
        "/plugin/api-keys": {
            "get": {
                "tags": ["Plugin"],
                "operationId": "GetApiKeys",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/ApiKeyEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            },
            "post": {
                "tags": ["Plugin"],
                "operationId": "CreateApiKey",
                "consumes": ["application/json-patch+json", "application/json", "text/json", "application/*+json"],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "obj", "in": "body", "required": false, "schema": { "$ref": "#/definitions/PluginConfigRequest" } }
                ],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/ApiKeyEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            }
        },
        "/plugin/api-keys/{apiKeyId}": {
            "get": {
                "tags": ["Plugin"],
                "operationId": "GetApiKey",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "apiKeyId", "in": "path", "required": true, "type": "string", "format": "uuid" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/ApiKeyEntity" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            },
            "delete": {
                "tags": ["Plugin"],
                "operationId": "DeleteApiKey",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "apiKeyId", "in": "path", "required": true, "type": "string", "format": "uuid" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/ActionResult" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/plugin/api-keys/{apiKeyId}/domains/{domainId}": {
            "delete": {
                "tags": ["Plugin"],
                "operationId": "DeleteApiKeyDomain",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "apiKeyId", "in": "path", "required": true, "type": "string", "format": "uuid" },
                    { "name": "domainId", "in": "path", "required": true, "type": "string", "format": "uuid" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/ActionResult" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/plugin/api-keys/{apiKeyid}/domains": {
            "post": {
                "tags": ["Plugin"],
                "operationId": "CreateApiKeyDomain",
                "consumes": ["application/json-patch+json", "application/json", "text/json", "application/*+json"],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "obj", "in": "body", "required": false, "schema": { "$ref": "#/definitions/PluginConfigRequest" } },
                    { "name": "apiKeyid", "in": "path", "required": true, "type": "string" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/DomainEntity" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/plugin/config": {
            "get": {
                "tags": ["Plugin"],
                "operationId": "GetConfig",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/ConfigEntity" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            },
            "put": {
                "tags": ["Plugin"],
                "operationId": "UpdateConfig",
                "consumes": ["application/json-patch+json", "application/json", "text/json", "application/*+json"],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "request", "in": "body", "required": false, "schema": { "$ref": "#/definitions/PluginConfigRequest" } }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "$ref": "#/definitions/ConfigEntity" } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "404": { "description": "Not Found" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/timezones": {
            "get": {
                "tags": ["TimeZone"],
                "operationId": "GetTimeZones",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/TimeZoneEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/timezones/{country}": {
            "get": {
                "tags": ["TimeZone"],
                "operationId": "GetTimeZonesByCountry",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "country", "in": "path", "required": true, "type": "string" }],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/TimeZoneEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        }
    },
    "definitions": {
...
        "ApiKeyEntity": {
            "type": "object",
            "properties": {
                "apiKeyId": { "format": "uuid", "type": "string" },
                "key": { "type": "string" },
                "expiresOn": { "format": "date-time", "type": "string" },
                "domains": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/DomainEntity" } }
            }
        },
        "DomainEntity": { "type": "object", "properties": { "domainId": { "format": "uuid", "type": "string" }, "name": { "type": "string" } } },
        "PluginConfigRequest": { "type": "object", "properties": { "config": { "$ref": "#/definitions/ConfigEntity" } } },
        "ConfigEntity": {
            "type": "object",
            "properties": {
                "configId": { "format": "uuid", "type": "string" },
                "name": { "type": "string" },
                "embedExternalImages": { "type": "boolean" },
                "removeEmptyTags": { "type": "boolean" },
                "removeSpanTags": { "type": "boolean" },
                "removeClassNames": { "type": "boolean" },
                "removeIframes": { "type": "boolean" },
                "removeTagAttributes": { "type": "boolean" }
            }
        },
        "ActionResult": { "type": "object", "properties": {} },
        "TimeZoneEntity": { "type": "object", "properties": { "name": { "type": "string" }, "displayName": { "type": "string" }, "offset": { "type": "string" } } }
    }
}